Refactor: Reorganize imports, move utils, and update context hooks
