+++ src/components/calendar/Calendar.tsx
@@ -18,7 +18,7 @@ import { cn } from '../../lib/utils';
 import { Task } from '../../types/task';
 import { GlassCard } from '../ui/GlassCard';
 import { WaveButton } from '../ui/WaveButton';
-import { Typography } from '../ui/typography';
+import { Typography } from '../ui/typography-utils';
 import { toDate } from '../../utils/dateHelpers';
 
 interface CalendarProps {
