+++ src/components/settings/components/SettingsNavigation.tsx
@@ -2,7 +2,7 @@ import { Bell, Download, Shield, User } from 'lucide-react';
 import React from 'react';
 import { cn } from '../../../lib/utils';
 import { WaveButton } from '../../ui/WaveButton';
-import { Typography } from '../../ui/typography';
+import { Typography } from '../../ui/typography-utils';
 import type { SettingsTab } from '../types';
 
 interface SettingsNavigationProps {
