+++ src/components/settings/sections/DataSection.tsx
@@ -9,12 +9,12 @@ import {
   Upload,
 } from 'lucide-react';
 import { useEffect, useState } from 'react';
-import { useAuth } from '../../../contexts/AuthContext';
+import { useAuth } from '../../../hooks/useAuth';
 import { BackupScheduler, BackupService } from '../../../lib/backup';
 import { cn } from '../../../lib/utils';
 import { InlineLoading } from '../../common/LoadingSpinner';
 import { WaveButton } from '../../ui/WaveButton';
-import { Typography } from '../../ui/typography';
+import { Typography } from '../../ui/typography-utils';
 import type { SettingsSectionProps } from '../types';
 
 export function DataSection({
