+++ src/components/settings/sections/NotificationSection.tsx
@@ -10,7 +10,7 @@ import {
 import { useEffect, useState } from 'react';
 import { cn } from '../../../lib/utils';
 import { WaveButton } from '../../ui/WaveButton';
-import { Typography } from '../../ui/typography';
+import { Typography } from '../../ui/typography-utils';
 import type { SettingsSectionProps } from '../types';
 
 export function NotificationSection({
