+++ src/components/settings/sections/PrivacySection.tsx
@@ -1,7 +1,7 @@
 import { BarChart3, Eye, EyeOff, Shield, Users } from 'lucide-react';
 import { cn } from '../../../lib/utils';
 import { WaveButton } from '../../ui/WaveButton';
-import { Typography } from '../../ui/typography';
+import { Typography } from '../../ui/typography-utils';
 import type { SettingsSectionProps } from '../types';
 
 export function PrivacySection({
