+++ src/components/statistics/StatisticsInsights.tsx
@@ -12,7 +12,7 @@ import {
 } from 'lucide-react';
 import { GlassCard } from '../ui/GlassCard';
 import { WaveButton } from '../ui/WaveButton';
-import { Typography } from '../ui/typography';
+import { Typography } from '../ui/typography-utils';
 import { LoadingSpinner } from '../common/LoadingSpinner';
 import { 
   statisticsAnalyzer,
