+++ src/components/task/TaskCommentSection.tsx
@@ -14,9 +14,9 @@ import {
   PopoverContent,
   PopoverTrigger,
 } from '@/components/ui/popover';
-import { Typography } from '@/components/ui/typography';
+import { Typography } from '@/components/ui/typography-utils';
 import { cn } from '@/components/ui/utils';
-import { useAuth } from '@/contexts/AuthContext';
+import { useAuth } from '@/hooks/useAuth';
 import { useData } from '@/contexts/DataContext';
 import { Comment, useComments } from '@/hooks/useComments';
 import { formatDistanceToNow } from 'date-fns';
