+++ src/components/ui/typography-utils.ts
@@ -1,0 +1,36 @@
+import {
+  H1,
+  H2,
+  H3,
+  H4,
+  H5,
+  H6,
+  BodyLarge,
+  Body,
+  BodySmall,
+  Caption,
+  Label,
+  Display,
+  Code,
+  Quote,
+} from './typography';
+
+// Typography object for easy importing
+export const Typography = {
+  H1,
+  H2,
+  H3,
+  H4,
+  H5,
+  H6,
+  BodyLarge,
+  Body,
+  BodySmall,
+  Caption,
+  Label,
+  Display,
+  Code,
+  Quote,
+};
+
+export default Typography;
