+++ src/components/ui/typography.tsx
@@ -250,22 +250,4 @@ export const Quote = ({
   </Component>
 );
 
-// Typography object for easy importing
-export const Typography = {
-  H1,
-  H2,
-  H3,
-  H4,
-  H5,
-  H6,
-  BodyLarge,
-  Body,
-  BodySmall,
-  Caption,
-  Label,
-  Display,
-  Code,
-  Quote,
-};
 
-export default Typography;
