+++ src/contexts/AppContext.tsx
@@ -260,16 +260,7 @@ export interface AppContextType {
 }
 
 // Create Context
-const AppContext = createContext<AppContextType | undefined>(undefined);
-
-// Hook to use App Context
-export function useApp(): AppContextType {
-  const context = useContext(AppContext);
-  if (context === undefined) {
-    throw new Error('useApp must be used within an AppProvider');
-  }
-  return context;
-}
+export const AppContext = createContext<AppContextType | undefined>(undefined);
 
 // App Provider Component
 export function AppProvider({ children }: { children: React.ReactNode }) {
@@ -355,7 +346,7 @@ export function AppProvider({ children }: { children: React.ReactNode }) {
             payload: { group: defaultGroup, groupId: defaultGroup.id },
           });
         }
-      } catch (error) {
+      } catch (_error) {
         dispatch({
           type: 'SET_ERROR',
           payload: {
