+++ src/contexts/TaskContext.tsx
@@ -2,7 +2,7 @@ import React, { createContext, useContext, useReducer, useEffect } from 'react';
 import { Task, TaskStatus, CreateTaskInput, UpdateTaskInput } from '../types/task';
 import { taskService } from '../lib/firestore';
 import { useAuth } from './AuthContext';
-import { useApp } from './AppContext';
+import { useApp } from '../hooks/useApp';
 
 // Task State Interface
 export interface TaskState {
