+++ src/hooks/useGroup.ts
@@ -1,5 +1,5 @@
 import { useCallback, useEffect, useState } from 'react';
-import { useAuth } from '../contexts/AuthContext';
+import { useAuth } from './useAuth';
 import { groupService } from '../lib/firestore';
 import {
   CreateGroupInput,
