+++ src/pages/Notifications.tsx
@@ -15,8 +15,8 @@ import { LoadingSpinner } from '../components/common/LoadingSpinner';
 import { WaveBackground } from '../components/layout/WaveBackground';
 import { GlassCard } from '../components/ui/GlassCard';
 import { WaveButton } from '../components/ui/WaveButton';
-import { Typography } from '../components/ui/typography';
-import { useAuth } from '../contexts/AuthContext';
+import { Typography } from '../components/ui/typography-utils';
+import { useAuth } from '../hooks/useAuth';
 import { NotificationService } from '../lib/notifications';
 import { Notification, NotificationStats } from '../types/notification';
 import { cn } from '../lib/utils';
