+++ src/pages/PointsManagement.tsx
@@ -22,8 +22,8 @@ import { TaskDetailModal } from '../components/points/TaskDetailModal';
 import { PointAnalysisModal } from '../components/points/PointAnalysisModal';
 import { GlassCard } from '../components/ui/GlassCard';
 import { WaveButton } from '../components/ui/WaveButton';
-import { Typography } from '../components/ui/typography';
-import { useAuth } from '../contexts/AuthContext';
+import { Typography } from '../components/ui/typography-utils';
+import { useAuth } from '../hooks/useAuth';
 import { useGroup, useUserGroups } from '../hooks/useGroup';
 import { enhancedUserService } from '../lib/firestore-improved';
 import { PointHistory, PointStats, pointsService } from '../lib/points';
