+++ src/pages/Statistics.tsx
@@ -18,10 +18,11 @@ import { LoadingSpinner } from '../components/common/LoadingSpinner';
 import { WaveBackground } from '../components/layout/WaveBackground';
 import { StatisticsInsights } from '../components/statistics/StatisticsInsights';
 import { GlassCard } from '../components/ui/GlassCard';
-import { AvatarWrapper, getAvatarInitials } from '../components/ui/avatar-utils';
-import { Typography } from '../components/ui/typography';
+import { AvatarWrapper } from '../components/ui/avatar-utils';
+import { getAvatarInitials } from '../utils/avatar-utils';
+import { Typography } from '../components/ui/typography-utils';
 import { WaveButton } from '../components/ui/WaveButton';
-import { useAuth } from '../contexts/AuthContext';
+import { useAuth } from '../hooks/useAuth';
 import { useGroup, useUserGroups } from '../hooks/useGroup';
 import { useTasks } from '../hooks/useTasks';
 import { FilterUtils } from '../lib/design-tokens';
